requirements.txt
click==8.1.8
dlib==19.24.8
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.16.1
fsspec==2025.3.0
imageio==2.35.1
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
mysql-connector-python==9.0.0
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
pillow==10.4.0
rawpy==0.21.0
sympy==1.13.3
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
typing_extensions==4.13.2
